core: Fix libdwarf and Qt build failure
libdwarf headers are in /usr/include/libdwarf (at least on Debian),
and the core code now uses Qt directly so the QtWidgets cflags/ldflags
are needed to build the core library. Not sure if the core should be
using Qt directly, but this gets it to build again on GNU/Linux.